LayoutTests/css3/font-variant-parsing.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body>
<script>
function parse(property, propertyText) {
    var element = document.createElement("span");
    element.setAttribute("id", "testElement");
    if (property != "")
        element.setAttribute("style", property + ": " + propertyText + ";");
    document.body.appendChild(element);
    let result = getComputedStyle(document.getElementById('testElement')).getPropertyValue(property)
    document.body.removeChild(element);
    return result
}

shouldBeEqualToString("parse('font-variant-ligatures', '')", "normal")
shouldBeEqualToString("parse('font-variant-ligatures', 'normal')", "normal")
shouldBeEqualToString("parse('font-variant-ligatures', 'normal common-ligatures')", "normal")
shouldBeEqualToString("parse('font-variant-ligatures', 'notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-ligatures', 'common-ligatures')", "common-ligatures")
shouldBeEqualToString("parse('font-variant-ligatures', 'common-ligatures common-ligatures')", "normal")
shouldBeEqualToString("parse('font-variant-ligatures', 'no-common-ligatures')", "no-common-ligatures")
shouldBeEqualToString("parse('font-variant-ligatures', 'discretionary-ligatures')", "discretionary-ligatures")
shouldBeEqualToString("parse('font-variant-ligatures', 'no-discretionary-ligatures')", "no-discretionary-ligatures")
shouldBeEqualToString("parse('font-variant-ligatures', 'historical-ligatures')", "historical-ligatures")
shouldBeEqualToString("parse('font-variant-ligatures', 'no-historical-ligatures')", "no-historical-ligatures")
shouldBeEqualToString("parse('font-variant-ligatures', 'contextual')", "contextual")
shouldBeEqualToString("parse('font-variant-ligatures', 'no-contextual')", "no-contextual")
shouldBeEqualToString("parse('font-variant-ligatures', 'common-ligatures no-contextual')", "common-ligatures no-contextual")
shouldBeEqualToString("parse('font-variant-ligatures', 'no-contextual common-ligatures')", "common-ligatures no-contextual")
shouldBeEqualToString("parse('font-variant-ligatures', 'no-contextual common-ligatures notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-ligatures', 'none')", "none")
shouldBeEqualToString("parse('font-variant-ligatures', 'no-common-ligatures no-discretionary-ligatures no-historical-ligatures no-contextual')", "none")

shouldBeEqualToString("parse('font-variant-position', '')", "normal")
shouldBeEqualToString("parse('font-variant-position', 'normal')", "normal")
shouldBeEqualToString("parse('font-variant-position', 'notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-position', 'sub')", "sub")
shouldBeEqualToString("parse('font-variant-position', 'super')", "super")
shouldBeEqualToString("parse('font-variant-position', 'sub super')", "normal")

shouldBeEqualToString("parse('font-variant-caps', '')", "normal")
shouldBeEqualToString("parse('font-variant-caps', 'normal')", "normal")
shouldBeEqualToString("parse('font-variant-caps', 'notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-caps', 'normal notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-caps', 'normal small-caps')", "normal")
shouldBeEqualToString("parse('font-variant-caps', 'small-caps')", "small-caps")
shouldBeEqualToString("parse('font-variant-caps', 'all-small-caps')", "all-small-caps")
shouldBeEqualToString("parse('font-variant-caps', 'petite-caps')", "petite-caps")
shouldBeEqualToString("parse('font-variant-caps', 'all-petite-caps')", "all-petite-caps")
shouldBeEqualToString("parse('font-variant-caps', 'unicase')", "unicase")
shouldBeEqualToString("parse('font-variant-caps', 'titling-caps')", "titling-caps")
shouldBeEqualToString("parse('font-variant-caps', 'unicase titling-caps')", "normal")
shouldBeEqualToString("parse('font-variant-caps', 'small-caps all-small-caps')", "normal")

shouldBeEqualToString("parse('font-variant-numeric', '')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'normal')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'normal notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'normal lining-nums')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'lining-nums')", "lining-nums")
shouldBeEqualToString("parse('font-variant-numeric', 'oldstyle-nums')", "oldstyle-nums")
shouldBeEqualToString("parse('font-variant-numeric', 'proportional-nums')", "proportional-nums")
shouldBeEqualToString("parse('font-variant-numeric', 'tabular-nums')", "tabular-nums")
shouldBeEqualToString("parse('font-variant-numeric', 'diagonal-fractions')", "diagonal-fractions")
shouldBeEqualToString("parse('font-variant-numeric', 'stacked-fractions')", "stacked-fractions")
shouldBeEqualToString("parse('font-variant-numeric', 'ordinal')", "ordinal")
shouldBeEqualToString("parse('font-variant-numeric', 'slashed-zero')", "slashed-zero")
shouldBeEqualToString("parse('font-variant-numeric', 'lining-nums slashed-zero')", "lining-nums slashed-zero")
shouldBeEqualToString("parse('font-variant-numeric', 'lining-nums oldstyle-nums')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'proportional-nums tabular-nums')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'diagonal-fractions stacked-fractions')", "normal")
shouldBeEqualToString("parse('font-variant-numeric', 'lining-nums proportional-nums')", "lining-nums proportional-nums")
shouldBeEqualToString("parse('font-variant-numeric', 'lining-nums diagonal-fractions')", "lining-nums diagonal-fractions")
shouldBeEqualToString("parse('font-variant-numeric', 'ordinal slashed-zero')", "ordinal slashed-zero")

shouldBeEqualToString("parse('font-variant-alternates', '')", "normal")
shouldBeEqualToString("parse('font-variant-alternates', 'normal')", "normal")
shouldBeEqualToString("parse('font-variant-alternates', 'notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-alternates', 'normal notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-alternates', 'historical-forms')", "historical-forms")
shouldBeEqualToString("parse('font-variant-alternates', 'normal historical-forms')", "normal")

shouldBeEqualToString("parse('font-variant-east-asian', '')", "normal")
shouldBeEqualToString("parse('font-variant-east-asian', 'normal')", "normal")
shouldBeEqualToString("parse('font-variant-east-asian', 'notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-east-asian', 'normal notavalidvalue')", "normal")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis78')", "jis78")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis83')", "jis83")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis90')", "jis90")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis04')", "jis04")
shouldBeEqualToString("parse('font-variant-east-asian', 'simplified')", "simplified")
shouldBeEqualToString("parse('font-variant-east-asian', 'traditional')", "traditional")
shouldBeEqualToString("parse('font-variant-east-asian', 'full-width')", "full-width")
shouldBeEqualToString("parse('font-variant-east-asian', 'proportional-width')", "proportional-width")
shouldBeEqualToString("parse('font-variant-east-asian', 'ruby')", "ruby")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis78 jis83')", "normal")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis90 traditional')", "normal")
shouldBeEqualToString("parse('font-variant-east-asian', 'full-width traditional')", "traditional full-width")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis04 proportional-width')", "jis04 proportional-width")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis04 proportional-width ruby')", "jis04 proportional-width ruby")
shouldBeEqualToString("parse('font-variant-east-asian', 'jis83 ruby')", "jis83 ruby")
shouldBeEqualToString("parse('font-variant-east-asian', 'full-width proportional-width')", "normal")
</script>
</body>
</html>
