LayoutTests/css3/font-variant-parsing-expected.txt

PASS parse('font-variant-ligatures', '') is "normal"
PASS parse('font-variant-ligatures', 'normal') is "normal"
PASS parse('font-variant-ligatures', 'normal common-ligatures') is "normal"
PASS parse('font-variant-ligatures', 'notavalidvalue') is "normal"
PASS parse('font-variant-ligatures', 'common-ligatures') is "common-ligatures"
PASS parse('font-variant-ligatures', 'common-ligatures common-ligatures') is "normal"
PASS parse('font-variant-ligatures', 'no-common-ligatures') is "no-common-ligatures"
PASS parse('font-variant-ligatures', 'discretionary-ligatures') is "discretionary-ligatures"
PASS parse('font-variant-ligatures', 'no-discretionary-ligatures') is "no-discretionary-ligatures"
PASS parse('font-variant-ligatures', 'historical-ligatures') is "historical-ligatures"
PASS parse('font-variant-ligatures', 'no-historical-ligatures') is "no-historical-ligatures"
PASS parse('font-variant-ligatures', 'contextual') is "contextual"
PASS parse('font-variant-ligatures', 'no-contextual') is "no-contextual"
PASS parse('font-variant-ligatures', 'common-ligatures no-contextual') is "common-ligatures no-contextual"
PASS parse('font-variant-ligatures', 'no-contextual common-ligatures') is "common-ligatures no-contextual"
PASS parse('font-variant-ligatures', 'no-contextual common-ligatures notavalidvalue') is "normal"
PASS parse('font-variant-ligatures', 'none') is "none"
PASS parse('font-variant-ligatures', 'no-common-ligatures no-discretionary-ligatures no-historical-ligatures no-contextual') is "none"
PASS parse('font-variant-position', '') is "normal"
PASS parse('font-variant-position', 'normal') is "normal"
PASS parse('font-variant-position', 'notavalidvalue') is "normal"
PASS parse('font-variant-position', 'sub') is "sub"
PASS parse('font-variant-position', 'super') is "super"
PASS parse('font-variant-position', 'sub super') is "normal"
PASS parse('font-variant-caps', '') is "normal"
PASS parse('font-variant-caps', 'normal') is "normal"
PASS parse('font-variant-caps', 'notavalidvalue') is "normal"
PASS parse('font-variant-caps', 'normal notavalidvalue') is "normal"
PASS parse('font-variant-caps', 'normal small-caps') is "normal"
PASS parse('font-variant-caps', 'small-caps') is "small-caps"
PASS parse('font-variant-caps', 'all-small-caps') is "all-small-caps"
PASS parse('font-variant-caps', 'petite-caps') is "petite-caps"
PASS parse('font-variant-caps', 'all-petite-caps') is "all-petite-caps"
PASS parse('font-variant-caps', 'unicase') is "unicase"
PASS parse('font-variant-caps', 'titling-caps') is "titling-caps"
PASS parse('font-variant-caps', 'unicase titling-caps') is "normal"
PASS parse('font-variant-caps', 'small-caps all-small-caps') is "normal"
PASS parse('font-variant-numeric', '') is "normal"
PASS parse('font-variant-numeric', 'normal') is "normal"
PASS parse('font-variant-numeric', 'notavalidvalue') is "normal"
PASS parse('font-variant-numeric', 'normal notavalidvalue') is "normal"
PASS parse('font-variant-numeric', 'normal lining-nums') is "normal"
PASS parse('font-variant-numeric', 'lining-nums') is "lining-nums"
PASS parse('font-variant-numeric', 'oldstyle-nums') is "oldstyle-nums"
PASS parse('font-variant-numeric', 'proportional-nums') is "proportional-nums"
PASS parse('font-variant-numeric', 'tabular-nums') is "tabular-nums"
PASS parse('font-variant-numeric', 'diagonal-fractions') is "diagonal-fractions"
PASS parse('font-variant-numeric', 'stacked-fractions') is "stacked-fractions"
PASS parse('font-variant-numeric', 'ordinal') is "ordinal"
PASS parse('font-variant-numeric', 'slashed-zero') is "slashed-zero"
PASS parse('font-variant-numeric', 'lining-nums slashed-zero') is "lining-nums slashed-zero"
PASS parse('font-variant-numeric', 'lining-nums oldstyle-nums') is "normal"
PASS parse('font-variant-numeric', 'proportional-nums tabular-nums') is "normal"
PASS parse('font-variant-numeric', 'diagonal-fractions stacked-fractions') is "normal"
PASS parse('font-variant-numeric', 'lining-nums proportional-nums') is "lining-nums proportional-nums"
PASS parse('font-variant-numeric', 'lining-nums diagonal-fractions') is "lining-nums diagonal-fractions"
PASS parse('font-variant-numeric', 'ordinal slashed-zero') is "ordinal slashed-zero"
PASS parse('font-variant-alternates', '') is "normal"
PASS parse('font-variant-alternates', 'normal') is "normal"
PASS parse('font-variant-alternates', 'notavalidvalue') is "normal"
PASS parse('font-variant-alternates', 'normal notavalidvalue') is "normal"
PASS parse('font-variant-alternates', 'historical-forms') is "historical-forms"
PASS parse('font-variant-alternates', 'normal historical-forms') is "normal"
PASS parse('font-variant-east-asian', '') is "normal"
PASS parse('font-variant-east-asian', 'normal') is "normal"
PASS parse('font-variant-east-asian', 'notavalidvalue') is "normal"
PASS parse('font-variant-east-asian', 'normal notavalidvalue') is "normal"
PASS parse('font-variant-east-asian', 'jis78') is "jis78"
PASS parse('font-variant-east-asian', 'jis83') is "jis83"
PASS parse('font-variant-east-asian', 'jis90') is "jis90"
PASS parse('font-variant-east-asian', 'jis04') is "jis04"
PASS parse('font-variant-east-asian', 'simplified') is "simplified"
PASS parse('font-variant-east-asian', 'traditional') is "traditional"
PASS parse('font-variant-east-asian', 'full-width') is "full-width"
PASS parse('font-variant-east-asian', 'proportional-width') is "proportional-width"
PASS parse('font-variant-east-asian', 'ruby') is "ruby"
PASS parse('font-variant-east-asian', 'jis78 jis83') is "normal"
PASS parse('font-variant-east-asian', 'jis90 traditional') is "normal"
PASS parse('font-variant-east-asian', 'full-width traditional') is "traditional full-width"
PASS parse('font-variant-east-asian', 'jis04 proportional-width') is "jis04 proportional-width"
PASS parse('font-variant-east-asian', 'jis04 proportional-width ruby') is "jis04 proportional-width ruby"
PASS parse('font-variant-east-asian', 'jis83 ruby') is "jis83 ruby"
PASS parse('font-variant-east-asian', 'full-width proportional-width') is "normal"
PASS successfullyParsed is true

TEST COMPLETE
